% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_asymmetry_2D.R
\name{detect_asymmetry_2D}
\alias{detect_asymmetry_2D}
\title{Title of the function}
\usage{
detect_asymmetry_2D(file_path, output_base = "asymmetry_output")
}
\arguments{
\item{file_path}{Path to input file}

\item{output_base}{Folder to store output}
}
\value{
A list containing results
}
\description{
Title of the function
}
